+++ backend/src/main/java/com/example/sellit/models/Role.java
@@ -19,7 +19,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_role", nullable = false)
-    private Integer idRole;
+    private Long idRole;
 
     @Column(name = "role_name", nullable = false)
     private String roleName;
